rename and move checkpoint activations related files
