+++ fairscale/nn/__init__.py
@@ -6,7 +6,8 @@
 from typing import List
 
 from .data_parallel import FullyShardedDataParallel, ShardedDataParallel
-from .misc import FlattenParamsWrapper, checkpoint_wrapper
+from .misc import FlattenParamsWrapper
+from .checkpoint import checkpoint_wrapper
 from .moe import MOELayer, Top2Gate
 from .pipe import Pipe, PipeRPCWrapper
 from .wrap import auto_wrap, default_auto_wrap_policy, enable_wrap, wrap
